@@ -12,7 +12,9 @@ using Datadog.Trace.ExtensionMethods;
 namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.Http.WebRequest
 {
     /// <summary>
-    /// CallTarget integration for HttpWebRequest.GetResponse
+    /// CallTarget integration for HttpWebRequest.GetResponse.
+    /// .NET 7+ uses an HttpClient internally (https://github.com/dotnet/runtime/pull/38511)
+    /// which is already instrumented.
     /// </summary>
     [InstrumentMethod(
         AssemblyName = WebRequestCommon.NetFrameworkAssembly,
