@@ -1,3 +1,3 @@
 #pragma once
 
-constexpr auto PROFILER_VERSION = "0.2.9";
+constexpr auto PROFILER_VERSION = "1.1.0";
