@@ -31,7 +31,7 @@ using static Datadog.Trace.ClrProfiler.AutoInstrumentation.AdoNet.AdoNetConstant
     AssemblyName = "System.Data.Common",
     TypeName = "System.Data.Common.DbCommand",
     MinimumVersion = "4.0.0",
-    MaximumVersion = "6.*.*",
+    MaximumVersion = "7.*.*",
     IntegrationName = nameof(IntegrationId.AdoNet),
     DataReaderType = TypeNames.DbDataReaderType,
     DataReaderTaskType = TypeNames.DbDataReaderTaskType,
