@@ -9,39 +9,43 @@ namespace Datadog.Trace.RuntimeMetrics
 {
     internal static class MetricsNames
     {
-        private const string MetricPrefix = "process.runtime.dotnet.";
-        public const string ExceptionsCount = $"{MetricPrefix}exceptions.count";
-
-        public const string ContentionTime = "runtime.dotnet.threads.contention_time";
-        public const string ContentionCount = $"{MetricPrefix}monitor.lock_contention.count";
-
-        public const string ThreadPoolWorkersCount = $"{MetricPrefix}thread_pool.threads.count";
-
-        public const string ThreadsCount = "runtime.dotnet.threads.count";
-
-        public const string CommittedMemory = "runtime.dotnet.mem.committed";
-
-        public const string CpuUserTime = "runtime.dotnet.cpu.user";
-        public const string CpuSystemTime = "runtime.dotnet.cpu.system";
-        public const string CpuPercentage = "runtime.dotnet.cpu.percent";
-
-        public const string AspNetCoreCurrentRequests = "runtime.dotnet.aspnetcore.requests.current";
-
-        public const string AspNetCoreFailedRequests = "runtime.dotnet.aspnetcore.requests.failed";
-        public const string AspNetCoreTotalRequests = "runtime.dotnet.aspnetcore.requests.total";
-        public const string AspNetCoreRequestQueueLength = "runtime.dotnet.aspnetcore.requests.queue_length";
-
-        public const string AspNetCoreCurrentConnections = "runtime.dotnet.aspnetcore.connections.current";
-        public const string AspNetCoreConnectionQueueLength = "runtime.dotnet.aspnetcore.connections.queue_length";
-        public const string AspNetCoreTotalConnections = "runtime.dotnet.aspnetcore.connections.total";
-
-        internal static class Gc
+        internal static class AspNetCore
         {
-            public const string CollectionsCount = $"{MetricPrefix}gc.collections.count";
-            public const string HeapSize = $"{MetricPrefix}gc.heap.size";
-            public const string AllocatedBytes = $"{MetricPrefix}gc.allocations.size";
-            public const string HeapCommittedMemory = $"{MetricPrefix}gc.committed_memory.size";
-            public const string PauseTime = $"{MetricPrefix}gc.pause.time";
+            public const string CoreCurrentRequests = "signalfx.dotnet.aspnetcore.requests.current";
+
+            public const string CoreFailedRequests = "signalfx.dotnet.aspnetcore.requests.failed";
+            public const string CoreTotalRequests = "signalfx.dotnet.aspnetcore.requests.total";
+            public const string CoreRequestQueueLength = "signalfx.dotnet.aspnetcore.requests.queue_length";
+
+            public const string CoreCurrentConnections = "signalfx.dotnet.aspnetcore.connections.current";
+            public const string CoreConnectionQueueLength = "signalfx.dotnet.aspnetcore.connections.queue_length";
+            public const string CoreTotalConnections = "signalfx.dotnet.aspnetcore.connections.total";
+        }
+
+        internal static class NetRuntime
+        {
+            public const string ExceptionsCount = "process.runtime.dotnet.exceptions.count";
+            public const string ContentionCount = "process.runtime.dotnet.monitor.lock_contention.count";
+            public const string ThreadPoolWorkersCount = "process.runtime.dotnet.thread_pool.threads.count";
+
+            internal static class Gc
+            {
+                public const string CollectionsCount = "process.runtime.dotnet.gc.collections.count";
+                public const string HeapSize = "process.runtime.dotnet.gc.heap.size";
+                public const string TotalObjectsSize = "process.runtime.dotnet.gc.objects.size";
+                public const string AllocatedBytes = "process.runtime.dotnet.gc.allocations.size";
+                public const string HeapCommittedMemory = "process.runtime.dotnet.gc.committed_memory.size";
+                public const string PauseTime = "process.runtime.dotnet.gc.pause.time";
+            }
+        }
+
+        internal static class Process
+        {
+            public const string MemoryUsage = "process.memory.usage";
+            public const string MemoryVirtual = "process.memory.virtual";
+            public const string CpuTime = "process.cpu.time";
+            public const string CpuUtilization = "process.cpu.utilization";
+            public const string ThreadsCount = "process.threads";
         }
     }
 }
