@@ -11,8 +11,9 @@ using System.Diagnostics;
 using System.Linq;
 using System.Threading.Tasks;
 using Datadog.Trace.Logging;
+using Datadog.Trace.SignalFx.Metrics;
 using Datadog.Trace.Util;
-using Datadog.Trace.Vendors.StatsdClient;
+using MetricType = Datadog.Tracer.SignalFx.Metrics.Protobuf.MetricType;
 
 namespace Datadog.Trace.RuntimeMetrics
 {
@@ -20,11 +21,11 @@ namespace Datadog.Trace.RuntimeMetrics
     {
         private const string MemoryCategoryName = ".NET CLR Memory";
         private const string ThreadingCategoryName = ".NET CLR LocksAndThreads";
-        private const string GarbageCollectionMetrics = $"{MetricsNames.Gc.HeapSize}, {MetricsNames.Gc.CollectionsCount}";
+        private const string GarbageCollectionMetrics = $"{MetricsNames.NetRuntime.Gc.HeapSize}, {MetricsNames.NetRuntime.Gc.CollectionsCount}";
 
         private static readonly IDatadogLogger Log = DatadogLogging.GetLoggerFor<PerformanceCountersListener>();
 
-        private readonly IDogStatsd _statsd;
+        private readonly ISignalFxMetricSender _metricSender;
         private readonly string _processName;
         private readonly int _processId;
 
@@ -40,9 +41,9 @@ namespace Datadog.Trace.RuntimeMetrics
 
         private Task _initializationTask;
 
-        public PerformanceCountersListener(IDogStatsd statsd)
+        public PerformanceCountersListener(ISignalFxMetricSender metricSender)
         {
-            _statsd = statsd;
+            _metricSender = metricSender;
 
             ProcessHelpers.GetCurrentProcessInformation(out _processName, out _, out _processId);
 
@@ -75,12 +76,12 @@ namespace Datadog.Trace.RuntimeMetrics
                 _instanceName = GetSimpleInstanceName();
             }
 
-            TryUpdateGauge(MetricsNames.Gc.HeapSize, _gen0Size, GcMetrics.Tags.Gen0);
-            TryUpdateGauge(MetricsNames.Gc.HeapSize, _gen1Size, GcMetrics.Tags.Gen1);
-            TryUpdateGauge(MetricsNames.Gc.HeapSize, _gen2Size, GcMetrics.Tags.Gen2);
-            TryUpdateGauge(MetricsNames.Gc.HeapSize, _lohSize, GcMetrics.Tags.LargeObjectHeap);
+            TryUpdateGauge(MetricsNames.NetRuntime.Gc.HeapSize, _gen0Size, GcMetrics.Tags.Gen0);
+            TryUpdateGauge(MetricsNames.NetRuntime.Gc.HeapSize, _gen1Size, GcMetrics.Tags.Gen1);
+            TryUpdateGauge(MetricsNames.NetRuntime.Gc.HeapSize, _gen2Size, GcMetrics.Tags.Gen2);
+            TryUpdateGauge(MetricsNames.NetRuntime.Gc.HeapSize, _lohSize, GcMetrics.Tags.LargeObjectHeap);
 
-            TryUpdateCounter(MetricsNames.ContentionCount, _contentionCount);
+            TryUpdateCounter(MetricsNames.NetRuntime.ContentionCount, _contentionCount);
 
             Log.Debug("Sent the following metrics: {metrics}", GarbageCollectionMetrics);
         }
@@ -122,7 +123,7 @@ namespace Datadog.Trace.RuntimeMetrics
 
             if (value != null)
             {
-                _statsd.Gauge(path, value.Value, tags: tags);
+                _metricSender.SendLong(path, value.Value, MetricType.GAUGE, tags);
             }
         }
 
@@ -135,7 +136,7 @@ namespace Datadog.Trace.RuntimeMetrics
                 return;
             }
 
-            _statsd.Counter(path, value.Value);
+            _metricSender.SendLong(path, value.Value, MetricType.CUMULATIVE_COUNTER);
         }
 
         private Tuple<string, bool> GetInstanceName()
