@@ -7,25 +7,37 @@
 
 using System;
 using System.Collections.Concurrent;
+using System.Collections.Generic;
+using System.Linq;
 using System.Runtime.ExceptionServices;
 using System.Threading;
+using Datadog.Trace.Configuration;
 using Datadog.Trace.Logging;
 using Datadog.Trace.PlatformHelpers;
+using Datadog.Trace.SignalFx.Metrics;
 using Datadog.Trace.Util;
+using Datadog.Trace.Vendors.Newtonsoft.Json.Utilities;
 using Datadog.Trace.Vendors.StatsdClient;
+using MetricType = Datadog.Tracer.SignalFx.Metrics.Protobuf.MetricType;
 
 namespace Datadog.Trace.RuntimeMetrics
 {
     internal class RuntimeMetricsWriter : IDisposable
     {
-        private const string ProcessMetrics = $"{MetricsNames.ThreadsCount}, {MetricsNames.CommittedMemory}, {MetricsNames.CpuUserTime}, {MetricsNames.CpuSystemTime}, {MetricsNames.CpuPercentage}";
+        private const string ProcessMetrics = $"{MetricsNames.Process.ThreadsCount}, {MetricsNames.Process.MemoryUsage}, {MetricsNames.Process.MemoryVirtual}, {MetricsNames.Process.CpuTime}, {MetricsNames.Process.CpuUtilization}";
+
+        private static readonly string[] CpuUserTag = new[] { "state:user" };
+        private static readonly string[] CpuSystemTag = new[] { "state:system" };
 
         private static readonly IDatadogLogger Log = DatadogLogging.GetLoggerFor<RuntimeMetricsWriter>();
-        private static readonly Func<IDogStatsd, TimeSpan, IRuntimeMetricsListener> InitializeListenerFunc = InitializeListener;
+        private static readonly Func<ImmutableMetricsIntegrationSettingsCollection, ISignalFxMetricSender, TimeSpan, IRuntimeMetricsListener> InitializeListenerFunc = InitializeListener;
 
+        private readonly IList<Action> sendMetricsActions = new List<Action>();
+
+        private readonly ImmutableMetricsIntegrationSettingsCollection _settings;
         private readonly TimeSpan _delay;
 
-        private readonly IDogStatsd _statsd;
+        private readonly ISignalFxMetricSender _metricSender;
         private readonly Timer _timer;
 
         private readonly IRuntimeMetricsListener _listener;
@@ -37,15 +49,21 @@ namespace Datadog.Trace.RuntimeMetrics
         private TimeSpan _previousUserCpu;
         private TimeSpan _previousSystemCpu;
 
-        public RuntimeMetricsWriter(IDogStatsd statsd, TimeSpan delay)
-            : this(statsd, delay, InitializeListenerFunc)
+        public RuntimeMetricsWriter(ImmutableMetricsIntegrationSettingsCollection settings, ISignalFxMetricSender metricSender, TimeSpan delay)
+            : this(settings, metricSender, delay, InitializeListenerFunc)
         {
         }
 
-        internal RuntimeMetricsWriter(IDogStatsd statsd, TimeSpan delay, Func<IDogStatsd, TimeSpan, IRuntimeMetricsListener> initializeListener)
+        internal RuntimeMetricsWriter(
+            ImmutableMetricsIntegrationSettingsCollection settings,
+            ISignalFxMetricSender metricSender,
+            TimeSpan delay,
+            Func<ImmutableMetricsIntegrationSettingsCollection, ISignalFxMetricSender, TimeSpan, IRuntimeMetricsListener> initializeListener)
         {
+            _settings = settings;
+
             _delay = delay;
-            _statsd = statsd;
+            _metricSender = metricSender;
             _timer = new Timer(_ => PushEvents(), null, delay, delay);
 
             try
@@ -59,7 +77,7 @@ namespace Datadog.Trace.RuntimeMetrics
 
             try
             {
-                ProcessHelpers.GetCurrentProcessRuntimeMetrics(out var userCpu, out var systemCpu, out _, out _);
+                ProcessHelpers.GetCurrentProcessRuntimeMetrics(out var userCpu, out var systemCpu, out _, out _, out _);
 
                 _previousUserCpu = userCpu;
                 _previousSystemCpu = systemCpu;
@@ -74,11 +92,25 @@ namespace Datadog.Trace.RuntimeMetrics
 
             try
             {
-                _listener = initializeListener(statsd, delay);
+                _listener = initializeListener(_settings, metricSender, delay);
             }
             catch (Exception ex)
             {
                 Log.Warning(ex, "Unable to initialize runtime listener, some runtime metrics will be missing");
+            }
+
+            var actionsForMetricsSending = new Dictionary<string, Action>
+            {
+                [MetricsIntegrationId.Process.ToString()] = SendProcessMetrics,
+                [MetricsIntegrationId.NetRuntime.ToString()] = SendNetRuntimeMetrics
+            };
+
+            foreach (var setting in settings.Settings)
+            {
+                if (setting.Enabled && actionsForMetricsSending.TryGetValue(setting.IntegrationName, out var action))
+                {
+                    sendMetricsActions.Add(action);
+                }
             }
         }
 
@@ -99,71 +131,96 @@ namespace Datadog.Trace.RuntimeMetrics
         {
             try
             {
-                _listener?.Refresh();
-
-                GcMetrics.PushCollectionCounts(_statsd);
-
-                if (_enableProcessMetrics)
+                foreach (var sendMetrics in sendMetricsActions)
                 {
-                    ProcessHelpers.GetCurrentProcessRuntimeMetrics(out var newUserCpu, out var newSystemCpu, out var threadCount, out var memoryUsage);
-
-                    var userCpu = newUserCpu - _previousUserCpu;
-                    var systemCpu = newSystemCpu - _previousSystemCpu;
-
-                    _previousUserCpu = newUserCpu;
-                    _previousSystemCpu = newSystemCpu;
-
-                    // Note: the behavior of Environment.ProcessorCount has changed a lot accross version: https://github.com/dotnet/runtime/issues/622
-                    // What we want is the number of cores attributed to the container, which is the behavior in 3.1.2+ (and, I believe, in 2.x)
-                    var maximumCpu = Environment.ProcessorCount * _delay.TotalMilliseconds;
-                    var totalCpu = userCpu + systemCpu;
-
-                    _statsd.Gauge(MetricsNames.ThreadsCount, threadCount);
-
-                    _statsd.Gauge(MetricsNames.CommittedMemory, memoryUsage);
-
-                    // Get CPU time in milliseconds per second
-                    _statsd.Gauge(MetricsNames.CpuUserTime, userCpu.TotalMilliseconds / _delay.TotalSeconds);
-                    _statsd.Gauge(MetricsNames.CpuSystemTime, systemCpu.TotalMilliseconds / _delay.TotalSeconds);
-
-                    _statsd.Gauge(MetricsNames.CpuPercentage, Math.Round(totalCpu.TotalMilliseconds * 100 / maximumCpu, 1, MidpointRounding.AwayFromZero));
-
-                    Log.Debug("Sent the following metrics: {metrics}", ProcessMetrics);
-                }
-
-                if (!_exceptionCounts.IsEmpty)
-                {
-                    foreach (var element in _exceptionCounts)
-                    {
-                        _statsd.Increment(MetricsNames.ExceptionsCount, element.Value, tags: new[] { $"exception_type:{element.Key}" });
-                    }
-
-                    // There's a race condition where we could clear items that haven't been pushed
-                    // Having an exact exception count is probably not worth the overhead required to fix it
-                    _exceptionCounts.Clear();
-
-                    Log.Debug("Sent the following metrics: {metrics}", MetricsNames.ExceptionsCount);
-                }
-                else
-                {
-                    Log.Debug("Did not send the following metrics: {metrics}", MetricsNames.ExceptionsCount);
+                    sendMetrics();
                 }
             }
             catch (Exception ex)
             {
-                Log.Warning(ex, "Error while updating runtime metrics");
+                Log.Warning(ex, "Error while updating metrics");
             }
         }
 
-        private static IRuntimeMetricsListener InitializeListener(IDogStatsd statsd, TimeSpan delay)
+        private static IRuntimeMetricsListener InitializeListener(ImmutableMetricsIntegrationSettingsCollection settings, ISignalFxMetricSender metricSender, TimeSpan delay)
         {
 #if NETCOREAPP
-            return new RuntimeEventListener(statsd, delay);
+            if (settings[MetricsIntegrationId.NetRuntime].Enabled || settings[MetricsIntegrationId.AspNetCore].Enabled)
+            {
+                return new RuntimeEventListener(settings, metricSender, delay);
+            }
+
+            return null;
 #elif NETFRAMEWORK
-            return AzureAppServices.Metadata.IsRelevant ? new AzureAppServicePerformanceCounters(statsd) : new PerformanceCountersListener(statsd);
+            if (settings[MetricsIntegrationId.NetRuntime].Enabled)
+            {
+                return AzureAppServices.Metadata.IsRelevant ? new AzureAppServicePerformanceCounters(metricSender) : new PerformanceCountersListener(metricSender);
+            }
+
+            return null;
 #else
             return null;
 #endif
+        }
+
+        private void SendNetRuntimeMetrics()
+        {
+            _listener?.Refresh();
+
+            GcMetrics.PushCollectionCounts(_metricSender);
+            _metricSender.SendLong(MetricsNames.NetRuntime.Gc.TotalObjectsSize, GC.GetTotalMemory(false), MetricType.GAUGE);
+
+            if (!_exceptionCounts.IsEmpty)
+            {
+                foreach (var element in _exceptionCounts)
+                {
+                    _metricSender.SendLong(MetricsNames.NetRuntime.ExceptionsCount, element.Value, MetricType.COUNTER, new[] { $"exception_type:{element.Key}" });
+                }
+
+                // There's a race condition where we could clear items that haven't been pushed
+                // Having an exact exception count is probably not worth the overhead required to fix it
+                _exceptionCounts.Clear();
+
+                Log.Debug("Sent the following metrics: {metrics}", MetricsNames.NetRuntime.ExceptionsCount);
+            }
+            else
+            {
+                Log.Debug("Did not send the following metrics: {metrics}", MetricsNames.NetRuntime.ExceptionsCount);
+            }
+        }
+
+        private void SendProcessMetrics()
+        {
+            if (!_enableProcessMetrics)
+            {
+                return;
+            }
+
+            ProcessHelpers.GetCurrentProcessRuntimeMetrics(out var newUserCpu, out var newSystemCpu, out var threadCount, out var virtualMemory, out var workingSet);
+
+            var userCpu = newUserCpu - _previousUserCpu;
+            var systemCpu = newSystemCpu - _previousSystemCpu;
+
+            _previousUserCpu = newUserCpu;
+            _previousSystemCpu = newSystemCpu;
+
+            _metricSender.SendLong(MetricsNames.Process.ThreadsCount, threadCount, MetricType.GAUGE);
+
+            _metricSender.SendLong(MetricsNames.Process.MemoryUsage, workingSet, MetricType.GAUGE);
+            _metricSender.SendLong(MetricsNames.Process.MemoryVirtual, virtualMemory, MetricType.GAUGE);
+
+            _metricSender.SendDouble(MetricsNames.Process.CpuTime, newUserCpu.TotalSeconds, MetricType.CUMULATIVE_COUNTER, CpuUserTag);
+            _metricSender.SendDouble(MetricsNames.Process.CpuTime, newSystemCpu.TotalSeconds, MetricType.CUMULATIVE_COUNTER, CpuSystemTag);
+
+            // Note: the behavior of Environment.ProcessorCount has changed a lot across version: https://github.com/dotnet/runtime/issues/622
+            // What we want is the number of cores attributed to the container, which is the behavior in 3.1.2+ (and, I believe, in 2.x)
+            var maximumCpu = Environment.ProcessorCount * _delay.TotalSeconds;
+
+            // https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/687#discussion_r995076259
+            _metricSender.SendDouble(MetricsNames.Process.CpuUtilization, Math.Min(userCpu.TotalSeconds / maximumCpu, 1D), MetricType.GAUGE, CpuUserTag);
+            _metricSender.SendDouble(MetricsNames.Process.CpuUtilization, Math.Min(systemCpu.TotalSeconds / maximumCpu, 1D), MetricType.GAUGE, CpuSystemTag);
+
+            Log.Debug("Sent the following metrics: {metrics}", ProcessMetrics);
         }
 
         private void FirstChanceException(object sender, FirstChanceExceptionEventArgs e)
