@@ -1,6 +1,7 @@
 ﻿using System;
 using Datadog.Trace.DogStatsd;
 using Datadog.Trace.Vendors.StatsdClient;
+using MetricType = Datadog.Tracer.SignalFx.Metrics.Protobuf.MetricType;
 
 namespace Datadog.Trace.SignalFx.Metrics
 {
@@ -11,9 +12,9 @@ namespace Datadog.Trace.SignalFx.Metrics
     /// </summary>
     internal class SignalFxStats : IDogStatsd
     {
-        private readonly SignalFxMetricSender _metricSender;
+        private readonly ISignalFxMetricSender _metricSender;
 
-        public SignalFxStats(SignalFxMetricSender metricSender)
+        public SignalFxStats(ISignalFxMetricSender metricSender)
         {
             _metricSender = metricSender ?? throw new ArgumentNullException(nameof(metricSender));
             // TODO splunk: consider supporting telemetry with our metric exporter
@@ -24,7 +25,7 @@ namespace Datadog.Trace.SignalFx.Metrics
 
         public void Dispose()
         {
-            _metricSender?.Dispose();
+            // nothing to dispose
         }
 
         public void Configure(StatsdConfig config)
@@ -34,38 +35,26 @@ namespace Datadog.Trace.SignalFx.Metrics
 
         public void Counter(string statName, long value, double sampleRate = 1, string[] tags = null)
         {
-            // we want to map to cumulative counter, leave the name to reduce required changes
-            _metricSender.SendCumulativeCounterMetric(statName, value, tags);
-        }
-
-        public void Increment(string statName, int value = 1, double sampleRate = 1, string[] tags = null)
-        {
-            // we want to map to counter, leave the name to reduce required changes
-            _metricSender.SendCounterMetric(statName, value, tags);
-        }
-
-        public void IncrementDouble(string statName, double value, double sampleRate = 1, string[] tags = null)
-        {
-            _metricSender.SendDoubleCounterMetric(statName, value, tags);
+            _metricSender.SendLong(statName, value, MetricType.CUMULATIVE_COUNTER, tags);
         }
 
         public void Decrement(string statName, int value = 1, double sampleRate = 1, params string[] tags)
         {
         }
 
-        public void Gauge(string statName, double value, double sampleRate = 1, string[] tags = null)
-        {
-            _metricSender.SendGaugeMetric(statName, value, tags);
-        }
-
         public void Timer(string statName, double value, double sampleRate = 1, string[] tags = null)
         {
-            _metricSender.SendGaugeMetric(statName, value, tags);
+            _metricSender.SendDouble(statName, value, MetricType.GAUGE, tags);
         }
 
         public void Event(string title, string text, string alertType = null, string aggregationKey = null, string sourceType = null, int? dateHappened = null, string priority = null, string hostname = null, string[] tags = null)
         {
             // noop
+        }
+
+        public void Gauge(string statName, double value, double sampleRate = 1, string[] tags = null)
+        {
+            _metricSender.SendDouble(statName, value, MetricType.GAUGE, tags);
         }
 
         public void Histogram(string statName, double value, double sampleRate = 1, string[] tags = null)
@@ -76,6 +65,11 @@ namespace Datadog.Trace.SignalFx.Metrics
         public void Distribution(string statName, double value, double sampleRate = 1, string[] tags = null)
         {
             // noop
+        }
+
+        public void Increment(string statName, int value = 1, double sampleRate = 1, string[] tags = null)
+        {
+            _metricSender.SendLong(statName, value, MetricType.COUNTER, tags);
         }
 
         public void Set<T>(string statName, T value, double sampleRate = 1, string[] tags = null)
