@@ -17,6 +17,6 @@ namespace Datadog.Trace
         /// </summary>
         public const ulong MaxTraceId = 9_223_372_036_854_775_807;
 
-        public static readonly string AssemblyVersion = "0.2.9.0";
+        public static readonly string AssemblyVersion = "1.1.0.0";
     }
 }
