@@ -3,7 +3,7 @@
 // This file was automatically generated by the UpdateVendors tool.
 //------------------------------------------------------------------------------
 using System;
-using System.Threading.Tasks;
+using System.Threading;
 
 namespace Datadog.Trace.Vendors.StatsdClient.Worker
 {
@@ -11,7 +11,7 @@ namespace Datadog.Trace.Vendors.StatsdClient.Worker
     {
         public void Wait(TimeSpan duration)
         {
-            Task.Delay(duration).Wait();
+            Thread.Sleep(duration);
         }
     }
 }
