@@ -93,7 +93,7 @@ namespace Datadog.Trace.ClrProfiler.IntegrationTests.CI
                         AssertTargetSpanEqual(targetSpan, "signalfx.tracing.library", "dotnet-tracing");
 
                         // check the SignalFx library version
-                        AssertTargetSpanEqual(targetSpan, "signalfx.tracing.version", "0.2.9.0");
+                        AssertTargetSpanEqual(targetSpan, "signalfx.tracing.version", "1.1.0.0");
 
                         // checks the origin tag
                         CheckOriginTag(targetSpan);
