@@ -13,9 +13,10 @@ using System.Collections.Generic;
 using System.Diagnostics.Tracing;
 using System.Threading;
 using Datadog.Trace.RuntimeMetrics;
-using Datadog.Trace.Vendors.StatsdClient;
+using Datadog.Trace.SignalFx.Metrics;
 using Moq;
 using Xunit;
+using MetricType = Datadog.Tracer.SignalFx.Metrics.Protobuf.MetricType;
 
 namespace Datadog.Trace.Tests.RuntimeMetrics
 {
@@ -26,31 +27,32 @@ namespace Datadog.Trace.Tests.RuntimeMetrics
         [Fact]
         public void PushEvents()
         {
-            var statsd = new Mock<IDogStatsd>();
+            var metricSender = new Mock<ISignalFxMetricSender>();
 
-            using var listener = new RuntimeEventListener(statsd.Object, TimeSpan.FromSeconds(10));
+            var settings = SettingsGenerator.Generate();
+            using var listener = new RuntimeEventListener(settings, metricSender.Object, TimeSpan.FromSeconds(10));
 
             listener.Refresh();
 
-            statsd.Verify(s => s.Gauge(MetricsNames.ContentionTime, It.IsAny<double>(), 1, null), Times.Once);
-            statsd.Verify(s => s.Counter(MetricsNames.ContentionCount, It.IsAny<long>(), 1, null), Times.Once);
-            statsd.Verify(s => s.Gauge(MetricsNames.ThreadPoolWorkersCount, It.IsAny<double>(), 1, null), Times.Once);
+            metricSender.Verify(s => s.SendLong(MetricsNames.NetRuntime.ContentionCount, It.IsAny<long>(), MetricType.CUMULATIVE_COUNTER, null), Times.Once);
+            metricSender.Verify(s => s.SendLong(MetricsNames.NetRuntime.ThreadPoolWorkersCount, It.IsAny<long>(), MetricType.GAUGE, null), Times.Once);
         }
 
         [Fact]
         public void MonitorGarbageCollections()
         {
-            var statsd = new Mock<IDogStatsd>();
+            var metricSender = new Mock<ISignalFxMetricSender>();
 
             var mutex = new ManualResetEventSlim();
 
             // TotalPauseTime is pushed on the GcRestartEnd event, which should be the last event for any GC
-            statsd.Setup(s => s.IncrementDouble(MetricsNames.Gc.PauseTime, It.IsAny<double>(), It.IsAny<double>(), It.IsAny<string[]>()))
+            metricSender.Setup(s => s.SendDouble(MetricsNames.NetRuntime.Gc.PauseTime, It.IsAny<double>(), MetricType.COUNTER, It.IsAny<string[]>()))
                 .Callback(() => mutex.Set());
 
-            using var listener = new RuntimeEventListener(statsd.Object, TimeSpan.FromSeconds(10));
+            var settings = SettingsGenerator.Generate();
+            using var listener = new RuntimeEventListener(settings, metricSender.Object, TimeSpan.FromSeconds(10));
 
-            statsd.Invocations.Clear();
+            metricSender.Invocations.Clear();
 
             mutex.Reset(); // In case a GC was triggered when creating the listener
 
@@ -63,17 +65,17 @@ namespace Datadog.Trace.Tests.RuntimeMetrics
                 throw new TimeoutException("Timed-out waiting for pause times to be reported.");
             }
 
-            statsd.Verify(s => s.Gauge(MetricsNames.Gc.HeapSize, It.IsAny<double>(), It.IsAny<double>(), new[] { "generation:gen0" }), Times.AtLeastOnce);
-            statsd.Verify(s => s.Gauge(MetricsNames.Gc.HeapSize, It.IsAny<double>(), It.IsAny<double>(), new[] { "generation:gen1" }), Times.AtLeastOnce);
-            statsd.Verify(s => s.Gauge(MetricsNames.Gc.HeapSize, It.IsAny<double>(), It.IsAny<double>(), new[] { "generation:gen2" }), Times.AtLeastOnce);
-            statsd.Verify(s => s.Gauge(MetricsNames.Gc.HeapSize, It.IsAny<double>(), It.IsAny<double>(), new[] { "generation:loh" }), Times.AtLeastOnce);
+            metricSender.Verify(s => s.SendLong(MetricsNames.NetRuntime.Gc.HeapSize, It.IsAny<long>(), MetricType.GAUGE, new[] { "generation:gen0" }), Times.AtLeastOnce);
+            metricSender.Verify(s => s.SendLong(MetricsNames.NetRuntime.Gc.HeapSize, It.IsAny<long>(), MetricType.GAUGE, new[] { "generation:gen1" }), Times.AtLeastOnce);
+            metricSender.Verify(s => s.SendLong(MetricsNames.NetRuntime.Gc.HeapSize, It.IsAny<long>(), MetricType.GAUGE, new[] { "generation:gen2" }), Times.AtLeastOnce);
+            metricSender.Verify(s => s.SendLong(MetricsNames.NetRuntime.Gc.HeapSize, It.IsAny<long>(), MetricType.GAUGE, new[] { "generation:loh" }), Times.AtLeastOnce);
 
-            statsd.Verify(s => s.Counter(MetricsNames.Gc.AllocatedBytes, It.IsAny<long>(), It.IsAny<double>(), It.IsAny<string[]>()), Times.AtLeastOnce);
-            statsd.Verify(s => s.IncrementDouble(MetricsNames.Gc.PauseTime, It.IsAny<double>(), It.IsAny<double>(), It.IsAny<string[]>()), Times.AtLeastOnce);
+            metricSender.Verify(s => s.SendLong(MetricsNames.NetRuntime.Gc.AllocatedBytes, It.IsAny<long>(), MetricType.CUMULATIVE_COUNTER, It.IsAny<string[]>()), Times.AtLeastOnce);
+            metricSender.Verify(s => s.SendDouble(MetricsNames.NetRuntime.Gc.PauseTime, It.IsAny<double>(), MetricType.COUNTER, It.IsAny<string[]>()), Times.AtLeastOnce);
 
 #if NET6_0_OR_GREATER
-            statsd.Verify(s => s.Gauge(MetricsNames.Gc.HeapSize, It.IsAny<double>(), It.IsAny<double>(), new[] { "generation:poh" }), Times.AtLeastOnce);
-            statsd.Verify(s => s.Gauge(MetricsNames.Gc.HeapCommittedMemory, It.IsAny<double>(), It.IsAny<double>(), It.IsAny<string[]>()), Times.AtLeastOnce);
+            metricSender.Verify(s => s.SendLong(MetricsNames.NetRuntime.Gc.HeapSize, It.IsAny<long>(), MetricType.GAUGE, new[] { "generation:poh" }), Times.AtLeastOnce);
+            metricSender.Verify(s => s.SendLong(MetricsNames.NetRuntime.Gc.HeapCommittedMemory, It.IsAny<long>(), MetricType.GAUGE, It.IsAny<string[]>()), Times.AtLeastOnce);
 #endif
         }
 
@@ -104,18 +106,20 @@ namespace Datadog.Trace.Tests.RuntimeMetrics
                 new PollingCounter("Dummy", eventSource, callback)
             };
 
-            var statsd = new Mock<IDogStatsd>();
-            using var listener = new RuntimeEventListener(statsd.Object, TimeSpan.FromSeconds(1));
+            var metricSender = new Mock<ISignalFxMetricSender>();
+            var settings = SettingsGenerator.Generate();
+
+            using var listener = new RuntimeEventListener(settings, metricSender.Object, TimeSpan.FromSeconds(1));
 
             // Wait for the counters to be refreshed
             mutex.Wait();
 
-            statsd.Verify(s => s.Gauge(MetricsNames.AspNetCoreCurrentRequests, 1.0, 1, null), Times.AtLeastOnce);
-            statsd.Verify(s => s.Gauge(MetricsNames.AspNetCoreFailedRequests, 2.0, 1, null), Times.AtLeastOnce);
-            statsd.Verify(s => s.Gauge(MetricsNames.AspNetCoreTotalRequests, 4.0, 1, null), Times.AtLeastOnce);
-            statsd.Verify(s => s.Gauge(MetricsNames.AspNetCoreRequestQueueLength, 8.0, 1, null), Times.AtLeastOnce);
-            statsd.Verify(s => s.Gauge(MetricsNames.AspNetCoreConnectionQueueLength, 16.0, 1, null), Times.AtLeastOnce);
-            statsd.Verify(s => s.Gauge(MetricsNames.AspNetCoreTotalConnections, 32.0, 1, null), Times.AtLeastOnce);
+            metricSender.Verify(s => s.SendDouble(MetricsNames.AspNetCore.CoreCurrentRequests, 1.0, MetricType.GAUGE, null), Times.AtLeastOnce);
+            metricSender.Verify(s => s.SendDouble(MetricsNames.AspNetCore.CoreFailedRequests, 2.0, MetricType.GAUGE, null), Times.AtLeastOnce);
+            metricSender.Verify(s => s.SendDouble(MetricsNames.AspNetCore.CoreTotalRequests, 4.0, MetricType.GAUGE, null), Times.AtLeastOnce);
+            metricSender.Verify(s => s.SendDouble(MetricsNames.AspNetCore.CoreRequestQueueLength, 8.0, MetricType.GAUGE, null), Times.AtLeastOnce);
+            metricSender.Verify(s => s.SendDouble(MetricsNames.AspNetCore.CoreConnectionQueueLength, 16.0, MetricType.GAUGE, null), Times.AtLeastOnce);
+            metricSender.Verify(s => s.SendDouble(MetricsNames.AspNetCore.CoreTotalConnections, 32.0, MetricType.GAUGE, null), Times.AtLeastOnce);
 
             foreach (var counter in counters)
             {
