@@ -2,6 +2,7 @@
 using Datadog.Trace.SignalFx.Metrics;
 using Datadog.Tracer.SignalFx.Metrics.Protobuf;
 using FluentAssertions;
+using Moq;
 using Xunit;
 
 namespace Datadog.Trace.Tests.SignalFx.Metrics
@@ -9,14 +10,14 @@ namespace Datadog.Trace.Tests.SignalFx.Metrics
     public class SignalFxMetricSenderTests
     {
         [Fact]
-        public void Gauge_metric_types_are_sent()
+        public void Gauge_double_metrics_are_sent()
         {
-            var reporter = new TestWriter();
-            var sender = new SignalFxMetricSender(reporter, new[] { "tag1:v1" });
+            var testWriter = new TestWriter();
+            var sender = new SignalFxMetricSender(new[] { "tag1:v1" }, Mock.Of<ISignalFxMetricExporter>(), 100, (_, _) => testWriter);
 
-            sender.SendGaugeMetric("test_metric", 1.11, tags: new[] { "tag2:v2" });
+            sender.SendDouble("test_metric", 1.11, MetricType.GAUGE, new[] { "tag2:v2" });
 
-            var dataPointUploadMessages = reporter.SentMessages;
+            var dataPointUploadMessages = testWriter.SentMessages;
             dataPointUploadMessages.Should().HaveCount(1);
             var dataPoint = dataPointUploadMessages[0];
 
@@ -27,14 +28,50 @@ namespace Datadog.Trace.Tests.SignalFx.Metrics
         }
 
         [Fact]
-        public void Counter_metric_types_are_sent()
+        public void Gauge_long_metrics_are_sent()
         {
-            var reporter = new TestWriter();
-            var sender = new SignalFxMetricSender(reporter, new[] { "tag1:v1" });
+            var testWriter = new TestWriter();
+            var sender = new SignalFxMetricSender(new[] { "tag1:v1" }, Mock.Of<ISignalFxMetricExporter>(), 100, (_, _) => testWriter);
 
-            sender.SendCounterMetric("test_metric", 1, tags: new[] { "tag2:v2" });
+            sender.SendLong("test_metric", 1, MetricType.GAUGE, new[] { "tag2:v2" });
 
-            var dataPointUploadMessages = reporter.SentMessages;
+            var dataPointUploadMessages = testWriter.SentMessages;
+            dataPointUploadMessages.Should().HaveCount(1);
+            var dataPoint = dataPointUploadMessages[0];
+
+            AssertCommon(dataPoint);
+
+            dataPoint.value.intValue.Should().Be(1);
+            dataPoint.metricType.Should().Be(MetricType.GAUGE);
+        }
+
+        [Fact]
+        public void Counter_double_metrics_are_sent()
+        {
+            var testWriter = new TestWriter();
+            var sender = new SignalFxMetricSender(new[] { "tag1:v1" }, Mock.Of<ISignalFxMetricExporter>(), 100, (_, _) => testWriter);
+
+            sender.SendDouble("test_metric", 1.11, MetricType.COUNTER, new[] { "tag2:v2" });
+
+            var dataPointUploadMessages = testWriter.SentMessages;
+            dataPointUploadMessages.Should().HaveCount(1);
+            var dataPoint = dataPointUploadMessages[0];
+
+            AssertCommon(dataPoint);
+
+            dataPoint.value.doubleValue.Should().Be(1.11);
+            dataPoint.metricType.Should().Be(MetricType.COUNTER);
+        }
+
+        [Fact]
+        public void Counter_long_metrics_are_sent()
+        {
+            var testWriter = new TestWriter();
+            var sender = new SignalFxMetricSender(new[] { "tag1:v1" }, Mock.Of<ISignalFxMetricExporter>(), 100, (_, _) => testWriter);
+
+            sender.SendLong("test_metric", 1, MetricType.COUNTER, new[] { "tag2:v2" });
+
+            var dataPointUploadMessages = testWriter.SentMessages;
             dataPointUploadMessages.Should().HaveCount(1);
             var dataPoint = dataPointUploadMessages[0];
 
@@ -45,14 +82,14 @@ namespace Datadog.Trace.Tests.SignalFx.Metrics
         }
 
         [Fact]
-        public void Cumulative_counter_metric_types_are_sent()
+        public void Cumulative_counter_metrics_are_sent()
         {
-            var reporter = new TestWriter();
-            var sender = new SignalFxMetricSender(reporter, new[] { "tag1:v1" });
+            var testWriter = new TestWriter();
+            var sender = new SignalFxMetricSender(new[] { "tag1:v1" }, Mock.Of<ISignalFxMetricExporter>(), 100, (_, _) => testWriter);
 
-            sender.SendCumulativeCounterMetric("test_metric", 1, tags: new[] { "tag2:v2" });
+            sender.SendLong("test_metric", 1, MetricType.CUMULATIVE_COUNTER, new[] { "tag2:v2" });
 
-            var dataPointUploadMessages = reporter.SentMessages;
+            var dataPointUploadMessages = testWriter.SentMessages;
             dataPointUploadMessages.Should().HaveCount(1);
             var dataPoint = dataPointUploadMessages[0];
 
